added option to keep compiled templates in memory
